requirements.txt
alembic==1.4.2
astroid==2.3.3
certifi==2019.11.28
chardet==3.0.4
click==7.1.1
Flask==1.1.1
Flask-Migrate==2.5.3
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
idna==2.9
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.1
lazy-object-proxy==1.4.3
Mako==1.1.2
MarkupSafe==1.1.1
mccabe==0.6.1
psycopg2-binary==2.8.4
PyJWT==1.7.1
pylint==2.4.4
python-dateutil==2.8.1
python-editor==1.0.4
rauth==0.7.3
requests==2.23.0
six==1.14.0
SQLAlchemy==1.3.15
urllib3==1.25.8
Werkzeug==1.0.0
wrapt==1.11.2
